+++ src/Plugins/SmartStore.GoogleAnalytics/GoogleAnalyticPlugin.cs
@@ -31,9 +31,16 @@ namespace SmartStore.GoogleAnalytics
         /// <returns>Widget zones</returns>
         public IList<string> GetWidgetZones()
         {
-            return !string.IsNullOrWhiteSpace(_googleAnalyticsSettings.WidgetZone)
-                       ? new List<string>() { _googleAnalyticsSettings.WidgetZone }
-                       : new List<string>() { "head_html_tag" };
+            var zones = new List<string>() { "head_html_tag", "mobile_head_html_tag" };
+            if(!string.IsNullOrWhiteSpace(_googleAnalyticsSettings.WidgetZone))
+            {
+                zones = new List<string>() { 
+                    _googleAnalyticsSettings.WidgetZone, 
+                    _googleAnalyticsSettings.WidgetZone == "head_html_tag" ? "mobile_head_html_tag" : "mobile_body_end_html_tag_after"
+                };
+            }
+
+            return zones;
         }
 
         /// <summary>